% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases_deaths_example.R
\docType{data}
\name{cases_deaths_example}
\alias{cases_deaths_example}
\title{Covid-19 cases and deaths example table}
\format{
A data frame with 100,000 rows and 17 columns.
}
\usage{
cases_deaths_example
}
\description{
Example table with Covid-19 cases and deaths, considering data from the Brazilian Health Ministry from February 23, 2022.
}
\keyword{datasets}
