% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_hospital_admissions_data.R
\name{fetch_hospital_admissions_data}
\alias{fetch_hospital_admissions_data}
\title{Fetch hospital admissions data from SIH}
\usage{
fetch_hospital_admissions_data(
  uf_acronym = "PI",
  year_start = 2020,
  year_end = 2021,
  month_start = 1,
  month_end = 12
)
}
\arguments{
\item{uf_acronym}{UF acronym. Defaults to PI.}

\item{year_start}{Start year. Numeric.}

\item{year_end}{End year. Numeric.}

\item{month_start}{Start month. Numeric.}

\item{month_end}{End month. Numeric.}
}
\value{
A tibble.
}
\description{
Fetch hospital admissions data from SIH
}
