% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_regions.R
\docType{data}
\name{health_regions}
\alias{health_regions}
\title{Health regions table}
\format{
A data frame with 5,570 rows and 2 variables:
\describe{
\item{cod_mun}{Municipality code from IBGE, with 6 digits.}
\item{cod_rs_450}{Health region code.}
}
}
\usage{
health_regions
}
\description{
Table of Brazilian municipalities and health regions codes, considering the 450 health regions version.
}
\keyword{datasets}
