% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_health_region.R
\docType{data}
\name{population_health_region}
\alias{population_health_region}
\title{Health regions population table}
\format{
A data frame with 450 rows and 2 variables:
\describe{
\item{cod_rs_450}{Health region coode.}
\item{pop2021}{Population estimative for 2021.}
}
}
\usage{
population_health_region
}
\description{
Table of health regions population estimatives with data from IBGE for 2021.
}
\keyword{datasets}
