% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_mun.R
\docType{data}
\name{population_mun}
\alias{population_mun}
\title{Municipalities population table}
\format{
A data frame with 5,570 rows and 2 variables:
\describe{
\item{cod_mun}{UF code from IBGE.}
\item{pop2021}{Population estimative for 2021.}
}
}
\usage{
population_mun
}
\description{
Table of municipalities population estimatives with data from IBGE for 2021.
}
\keyword{datasets}
