% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_uf.R
\docType{data}
\name{population_uf}
\alias{population_uf}
\title{Population UF table}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{uf_code}{UF code from IBGE.}
\item{uf}{UF acronym.}
\item{pop2021}{Population estimative for 2021.}
}
}
\usage{
population_uf
}
\description{
Table of UF population estimatives with data from IBGE for 2021.
}
\keyword{datasets}
