% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_and_export_data_csv.R
\name{prepare_and_export_data_csv}
\alias{prepare_and_export_data_csv}
\title{Prepare and export data to CSV files}
\usage{
prepare_and_export_data_csv(
  cases_deaths_folder = "data-raw/cases_deaths_example/",
  dest_folder,
  year_start = 2021,
  year_end = 2021,
  month_start = 1,
  month_end = 1
)
}
\arguments{
\item{cases_deaths_folder}{Folder with cases and deaths files.}

\item{dest_folder}{Destination folder for CSV files.}

\item{year_start}{Start year. Numeric. For SIH data.}

\item{year_end}{End year. Numeric. For SIH data.}

\item{month_start}{Start month. Numeric. For SIH data.}

\item{month_end}{End month. Numeric. For SIH data.}
}
\value{
A tibble.
}
\description{
Prepare and export data to CSV files
}
