% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cases_deaths_health_regions.R
\name{prepare_cases_deaths_health_regions}
\alias{prepare_cases_deaths_health_regions}
\title{Prepare a tibble with cases and deaths for health regions}
\usage{
prepare_cases_deaths_health_regions(cases_deaths, uf_code = 22)
}
\arguments{
\item{cases_deaths}{A tibble from read_cases_deaths}

\item{uf_code}{UF code. Defaults to 22.}
}
\value{
A tibble.
}
\description{
Prepare a tibble with cases and deaths for health regions
}
