% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_cases_deaths_uf.R
\name{prepare_cases_deaths_uf}
\alias{prepare_cases_deaths_uf}
\title{Prepare a tibble with cases and deaths for the UF}
\usage{
prepare_cases_deaths_uf(
  cases_deaths,
  uf_code = 22,
  with_rate = TRUE,
  constant = 1e+05,
  digits = 2
)
}
\arguments{
\item{cases_deaths}{A tibble from read_cases_deaths}

\item{uf_code}{UF code. Defaults to 22.}

\item{with_rate}{Logical. Calculate per capita rate.}

\item{constant}{Numerical. Constant for rate.}

\item{digits}{Numerical. Digits for rates.}
}
\value{
A tibble.
}
\description{
Prepare a tibble with cases and deaths for the UF
}
