% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hospital_admissions_municipalities.R
\name{prepare_hospital_admissions_municipalities}
\alias{prepare_hospital_admissions_municipalities}
\title{Prepare a tibble with hospital admissions count for municipalities}
\usage{
prepare_hospital_admissions_municipalities(
  hospital_admissions_data,
  uf_code = 22
)
}
\arguments{
\item{hospital_admissions_data}{Data from \code{fetch_hospital_admissions_data} function.}

\item{uf_code}{UF code. Defaults to 22.}
}
\value{
A tibble.
}
\description{
Prepare a tibble with hospital admissions count for municipalities
}
