% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vaccination_uf.R
\name{prepare_vaccination}
\alias{prepare_vaccination}
\title{Prepare a tibble with vaccination counts for a UF}
\usage{
prepare_vaccination(lazy_sipni)
}
\arguments{
\item{lazy_sipni}{Lazi table from \code{read_sipni} function.}

\item{uf_acronym}{UF acronym. Defaults to PI}
}
\value{
A tibble.
}
\description{
Prepare a tibble with vaccination counts for a UF
}
