% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_vaccination_municipalities.R
\name{prepare_vaccination_municipalities}
\alias{prepare_vaccination_municipalities}
\title{Prepare a tibble with vaccination counts for municipalities within a UF.}
\usage{
prepare_vaccination_municipalities(lazy_sipni, uf_acronym = "PI")
}
\arguments{
\item{lazy_sipni}{Lazi table from \code{read_sipni} function.}

\item{uf_acronym}{UF acronym. Defaults to PI}
}
\value{
A tibble.
}
\description{
Prepare a tibble with vaccination counts for municipalities within a UF.
}
