% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cases_deaths.R
\name{read_cases_deaths}
\alias{read_cases_deaths}
\title{Read cases and deaths files from Brazilian Health Ministry}
\usage{
read_cases_deaths(files_folder)
}
\arguments{
\item{files_folder}{Path to folder containing CSV files from Braziliann Health Ministry}
}
\value{
A tibble.
}
\description{
Read cases and deaths files from Brazilian Health Ministry
}
