% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sivep_gripe.R
\name{read_sivep_gripe}
\alias{read_sivep_gripe}
\title{Read Sivep Gripe files from Brazilian Health Ministry}
\usage{
read_sivep_gripe(files_folder, chunk_size = 1e+05)
}
\arguments{
\item{files_folder}{Path to folder containing CSV files from Brazilian Health Ministry}

\item{chunk_size}{Integer. Chunk size to read CSV files. Default 100,000 lines.}
}
\value{
A lazy connection to Sivep Gripe table in a SQLite database stored in a temporary folder.
}
\description{
Read Sivep Gripe files from Brazilian Health Ministry
}
