% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sivep_gripe_header.R
\docType{data}
\name{sivep_gripe_header}
\alias{sivep_gripe_header}
\title{Sivep Gripe header}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 0 rows and 166 columns.
}
\usage{
sivep_gripe_header
}
\description{
Tibble with header, empty lines. Internal use.
}
\keyword{datasets}
