% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_hospital_admissions_uf.R
\name{prepare_hospital_admissions_uf}
\alias{prepare_hospital_admissions_uf}
\title{Prepare a tibble with health admissions count for a UF}
\usage{
prepare_hospital_admissions_uf(hospital_admissions_data, uf_code = 22)
}
\arguments{
\item{hospital_admissions_data}{Data from \code{fetch_hospital_admissions_data} function.}

\item{uf_code}{UF code. Defaults to 22.}
}
\value{
A tibble.
}
\description{
Prepare a tibble with health admissions count for a UF
}
